task pane working
