+++ Samples/excel-shared-runtime-global-state/.gitignore
@@ -1,0 +1,68 @@
+# Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
+# See LICENSE in the project root for license information.
+
+# Created by https://www.gitignore.io/api/osx,windows
+
+### OSX ###
+.DS_Store
+.AppleDouble
+.LSOverride
+
+# Icon must end with two \r
+Icon
+
+
+# Thumbnails
+._*
+
+# Files that might appear in the root of a volume
+.DocumentRevisions-V100
+.fseventsd
+.Spotlight-V100
+.TemporaryItems
+.Trashes
+.VolumeIcon.icns
+
+# Directories potentially created on remote AFP share
+.AppleDB
+.AppleDesktop
+Network Trash Folder
+Temporary Items
+.apdisk
+
+
+### Windows ###
+# Windows image file caches
+Thumbs.db
+ehthumbs.db
+
+# Folder config file
+Desktop.ini
+
+# Recycle Bin used on file shares
+$RECYCLE.BIN/
+
+# Windows Installer files
+*.cab
+*.msi
+*.msm
+*.msp
+
+# Windows shortcuts
+*.lnk
+
+# Ignore the following folders
+.vs
+.idea
+*.log
+obj
+bin
+*.user
+
+node_modules/**
+packages/**
+typings/**
+.vscode/**
+.awcache/**
+
+dist/**
