+++ Samples/excel-shared-runtime-global-state/src/functions/functions.ts
@@ -1,4 +1,4 @@
-﻿import { getGlobal } from "../commands/commands";
+﻿import { getValueForKey, setValueForKey } from "../taskpane/helpers";
 
 /**
  * Get value for key
@@ -28,28 +28,3 @@ export function setValueForKeyCF(key: string, value: string): string {
  * @param second Second number
  * @returns The sum of the two numbers.
  */
-
-export function setValueForKey(key: string, value: string) {
-  let g = getGlobal() as any;
-  if (g.state.storageType === "globalvar") {
-    g.state.keys.push(key);
-    g.state.values.push(value);
-  } else {
-    g.window.localStorage.setItem(key, value);
-  }
-}
-
-export function getValueForKey(key: string): string {
-  let g = getGlobal() as any;
-  let answer = "";
-  if (g.state.storageType === "globalvar") {
-    g.state.keys.forEach((element, index) => {
-      if (element === key) {
-        answer = g.state.values[index];
-      }
-    });
-  } else {
-    answer = g.window.localStorage.getItem(key);
-  }
-  return answer;
-}
